ImageJVirtualStack: move ImageProcessor creation to separate class
