+++ src/test/java/net/imglib2/img/display/imagej/ImageJVirtualStackTest.java
@@ -54,9 +54,8 @@ import org.junit.Test;
 
 import static org.junit.Assert.assertArrayEquals;
 import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertTrue;
-
 import static org.junit.Assert.assertNotSame;
+import static org.junit.Assert.assertTrue;
 
 public class ImageJVirtualStackTest
 {
